Refactoring: splitting the Cacheable and compacting features from Database and moving it to 'fhirpath' module (with the related changes).
