+++ fhir-server/src/test/java/au/csiro/pathling/fhir/EntityTagInterceptorTest.java
@@ -28,7 +28,7 @@ import static org.mockito.Mockito.when;
 import au.csiro.pathling.caching.EntityTagInterceptor;
 import au.csiro.pathling.config.HttpServerCachingConfiguration;
 import au.csiro.pathling.config.ServerConfiguration;
-import au.csiro.pathling.io.Database;
+import au.csiro.pathling.io.DatabaseComponent;
 import ca.uhn.fhir.rest.api.server.RequestDetails;
 import ca.uhn.fhir.rest.server.exceptions.NotModifiedException;
 import java.util.List;
@@ -49,7 +49,7 @@ class EntityTagInterceptorTest {
 
   static final String TAG = "abc123";
 
-  Database database;
+  DatabaseComponent database;
   ConformanceProvider conformanceProvider;
   HttpServletRequest request;
   RequestDetails requestDetails;
@@ -58,7 +58,7 @@ class EntityTagInterceptorTest {
 
   @BeforeEach
   void setUp() {
-    database = mock(Database.class);
+    database = mock(DatabaseComponent.class);
     conformanceProvider = mock(ConformanceProvider.class);
     request = mock(HttpServletRequest.class);
     requestDetails = mock(RequestDetails.class);
