+++ fhir-server/src/test/java/au/csiro/pathling/test/integration/AsyncTest.java
@@ -22,6 +22,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import au.csiro.pathling.io.Database;
+import au.csiro.pathling.io.DatabaseComponent;
 import au.csiro.pathling.test.helpers.TestHelpers;
 import java.net.MalformedURLException;
 import java.net.URI;
@@ -60,7 +61,7 @@ class AsyncTest extends IntegrationTest {
   SparkSession spark;
 
   @MockBean
-  Database database;
+  DatabaseComponent database;
 
   @LocalServerPort
   int port;
