+++ fhir-server/src/test/java/au/csiro/pathling/test/integration/ExtractTest.java
@@ -22,7 +22,7 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import au.csiro.pathling.io.Database;
+import au.csiro.pathling.io.DatabaseComponent;
 import au.csiro.pathling.test.helpers.TestHelpers;
 import ca.uhn.fhir.parser.IParser;
 import java.io.IOException;
@@ -60,7 +60,7 @@ class ExtractTest extends IntegrationTest {
   IParser jsonParser;
 
   @MockBean
-  Database database;
+  DatabaseComponent database;
 
   @LocalServerPort
   int port;
